# docs/swagger-ui/swagger.yaml
swagger: '2.0'
info:
  title: Phoenix price API - gRPC Gateway docs
  description: A REST interface for price and trade queries
  version: 1.0.0
paths:
...
  /activities/{address}:
    get:
      operationId: Activities
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              activities:
                type: array
                items:
                  type: object
                  properties:
                    user:
                      type: string
                    actionType:
                      type: string
                    baseCurrency:
                      type: string
                    baseVolume:
                      type: number
                      format: double
                    targetCurrency:
                      type: string
                    targetVolume:
                      type: number
                      format: double
                    timeStamp:
                      type: string
                      format: uint64
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    '@type':
                      type: string
                  additionalProperties: {}
      parameters:
        - name: address
          in: path
          required: true
          type: string
        - name: from
          in: query
          required: false
          type: string
          format: uint64
        - name: to
          in: query
          required: false
          type: string
          format: uint64
        - name: page
          in: query
          required: false
          type: integer
          format: int32
        - name: pageSize
          in: query
          required: false
          type: integer
          format: int32
      tags:
        - UserQuery
  /users/total:
    get:
      operationId: TotalUsers
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              totalUsers:
                type: string
                format: int64
              usersLast24h:
                type: string
                format: int64
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    '@type':
                      type: string
                  additionalProperties: {}
      tags:
        - UserQuery
...